chore: Use cli-runtime module to construct ClientConfig and support standard kubectl options such as --context, --as, --cluster, etc.'
