--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/aquasecurity/kubectl-who-can/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
