+++ pkg/cmd/list_test.go
