+++ pkg/cmd/namespace_validator_test.go
@@ -1,4 +1,4 @@
-package whocan
+package cmd
 
 import (
 	"errors"
