+++ pkg/cmd/resource_resolver.go
@@ -1,4 +1,4 @@
-package whocan
+package cmd
 
 import (
 	"fmt"
