+++ pkg/cmd/resource_resolver_test.go
@@ -1,4 +1,4 @@
-package whocan
+package cmd
 
 import (
 	"errors"
